refactor: delete BuyerProfit
