+++ src/test/java/buyer/BuyerResultTest.java
@@ -1,10 +1,15 @@
 package buyer;
 
+import lotto.Lotto;
 import lotto.Rank;
+import lotto.WinningLotto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
 import java.util.EnumMap;
+import java.util.List;
 import java.util.Map;
 
 public class BuyerResultTest {
@@ -18,5 +23,19 @@ public class BuyerResultTest {
         buyerResult.matches(Rank.FIFTH);
         //then
         Assertions.assertThat(buyerResult).isEqualTo(new BuyerResult(cpMap));
+    }
+
+    @DisplayName("수익률 테스트")
+    @Test
+    void profitTest() {
+        //given
+        List<Lotto> lotteries = new ArrayList<>();
+        lotteries.add(new Lotto(List.of(1, 2, 3, 4, 5, 6)));
+        WinningLotto winningLotto = new WinningLotto(List.of(1, 2, 3, 4, 5, 6), 8);
+        //when
+        BuyerResult buyerResult = winningLotto.getResult(lotteries);
+        //then
+        Assertions.assertThat(buyerResult.getProfit()).isEqualTo((double) Rank.FIRST.prize / 1000);
+
     }
 }
